fix(prt): fix vertical ternary method (#2281)
Fix two bugs in the disv tracking method, one introduced in #1874, the other in #2183.

In the first case the logic to determine whether a particle exits a cell through a vertical or lateral face was incorrectly rearranged. This lost the original (correct) behavior for comparing exit times and selecting an exit face — where a particle should exit through the top or bottom, it could instead exit through a side.

In the second case the logic to "clamp" particles to the cell's vertical extent was incorrect, and particles already within that extent were moved abruptly to the bottom of the cell.

Both of these were probably missed up to now because the PRT test cases are currently mostly 1-layer. While a multi-layer test is added here, tt would be good to exercise PRT on more DISV models with depth.