docs(rtd): introduce migration guide page (#2287)
The MF6.6.0 PRT migration guide was shown in the API docs site rather than the RTD, because it's located in the src/ directory. Introduce a new page on the RTD site for version-specific migration guides and add a step copying the guide to the docs directory before building the RTD site. We currently do this for several files scattered throughout the repo — maybe eventually we can consolidate all or most of these in the docs folder. In the longer term, we have discussed consolidating our RTD and PDF docs. Maybe at that point migration guides can become part of the release notes PDF too.

Also add the extended build docs to the developer section, and fix a lingering reference to the old organization name.