refactor(dfn2f90.py): accept dfn files or directories (#2292)
Instead of accepting a text file specifying DFN files — a temporary measure while IDM supports a subset of the full specification — just accept one or more path arguments, each of which can be a DFN file or a folder containing DFN files, and feed the dfns.txt file to this from the shell.

A prep step before code generation is switched to consume TOML DFNs, coming shortly.

Note to model developers: after this, the behavior of running dfn2f90.py directly with no arguments will change to include all DFNs by default. To get the old behavior (i.e. the subset of DFNs IDM supports), it will be necessary to use the pixi task or feed dfns.txt to the script with e.g.

cat dfns.txt | xargs python ...