chore(dependencies): add tomli_w to pixi environment (#2298)
tomli_w is needed by the devtools models API as of MODFLOW-ORG/modflow-devtools#210. It'd get installed transitively if we used devtools optional dependency groups as part of the custom install task, but it's probably worth having it explicitly in the environment since we are adopting TOML more generally as a configuration format.