docs(prt): fix ref to removed option, restore dfns (#2309)
We had a lingering reference to TRACK_TIMES and TRACK_TIMESFILE, fix it. And also, I'm not sure why I removed the options from the DFNs in bf47ec6, for deprecations/removals we conventionally leave the options in the DFN and just mark them as such. The automation will then put them in the deprecation section in the release notes. Also fix the automation options that are removed without a deprecation period show the same version number for both deprecated and removed columns.