test(framework): allow access to local external models (#2300)
Support remote or local model access in the external model tests via an identical pattern. This uses some new stuff in the models API. This is relevant for anyone needing to modify test models during MF6 development.

By default, as has been the case since we started using the models API recently, the tests will use the GitHub-backed Pooch registry. To test against local models, introduce a --models-path pytest CLI option. If one or more is provided, the paths will be searched for models and tests will be run against those instead of models from the official registry.

By default, only MF6 models are discovered. There is a --namefile-pattern option to relax the discovery for MF2005 etc.

E.g. to run MF6 models in a local checkout of the modflow6-testmodels repo, from <project root>/autotest/ of the MODFLOW 6 repo

    pytest -m "external" --models-path ../../path/to/modflow6-testmodels/
    # or pixi run autotest -m ...

Or to run mf5to6 converter tests against the mf2005 models

    pytest -m "external" --models-path ../../path/to/modflow6-testmodels/mf5to6 --namefile-pattern "*.nam"

This PR updates the developer docs to reflect the above. Previously they described the old fixtures pattern.

Also update the CI to checkout the external model repos and use the new local registry mechanism rather than using the official registry. Since moving to the models API we have tested the models as frozen by the official registry but we want to test the latest. (Speaking of which, we will need to move to the develop/master branching strategy in test models and large test models repos so the official registry has something stable to point at — or consider moving those models into this repo.)

Also consolidate the external model tests from

    test_testmodels_mf6.py
    test_testmodels_mf5to6.py
    test_largetestmodels.py

into a single file test_external_models.py.