refactor(prt): introduce particle event system (#2376)
Building on #2373. Simple synchronous pub/sub.

Main motivation here is that prt needs to report things in many places throughout the solution instead of just one place like other models. This is more readable imo and cleans up some particle state transition code that was duplicated before. Tracking methods also no longer need to know about output files, just call the dispatcher to publish events.

We are paying for an allocation and a little extra overhead every time something happens to a particle, as opposed to just a procedure call before. But this seems worth it?

The dispatcher is responsible for the event lifecycle i.e. allocating before publishing, deallocating after consumption. If we want to batch events later on we will need to think on how to deallocate them on flush.

Only one consumer per dispatcher right now. This can be generalized to multiple consumers if needed.

I added a developer option DEV_DUMP_EVENT_TRACE to the output control package which prints verbose event summaries to stdout.

This is a convenience for debugging which will show all events, not just the ones selected for output to track files. The "context" is a string I stuck on the event type as another debugging aid, currently unused, maybe overkill.

Some more cleanup and reorganization too.

On review, Alden and I discussed defining event routines in the relevant methods, rather than in the dispatcher. This seems more appropriate because events might be scoped to a particular method type, while the plumbing should be able to handle any event type. Will come back to this.