test(prt): fix flowface/stopzone/extended tracking combo test (#2379)
Fix the test_prt_iflowface_istopzone.py autotest configuration to make sure PRT and MP7 are consuming identical flows and running with identical configurations (there were several differences). With this, endpoints now agree. Still some slightly different results with extended tracking on, but the differences seem small enough to chalk up to compiler toolchain and/or optimization flags. This makes me fairly confident we can close #2358.

But this does point to the fact that it's not easy to a) convert an MP7 simulation to an identical PRT simulation or b) verify that they are identically configured after doing so, without running them and comparing results. We need to consider how to improve flopy in this area.

Also, fix the fix introduced in #2369 for termination in cells with zero flow under extended tracking. It should only apply if extended tracking is on. This deserves more consideration (can we avoid special casing terminations like this at the subcell level, while guaranteeing that tracking time is accurately reported?) in followup.

Also, remove the context string on the new particle event type introduced recently. YAGNI