test: solve some memory leaks in the PtrHashTable unit tests (#2377)
When running Tester through valgrind some memoryleaks popped up in the (Ptr)HashTable related unit tests. The memoryleaks are all in the unit tests themself and not in the production code.

Still it is good to have the Tester memory leak free to not obfuscate any real memory leaks that might pop up in the future