refactor: introduce binary file reader module and base type (#2384)
Introduce an abstract base type with interface for binary file reading. Also factor out a routine for peeking at the next record's stress period and time step.

Preliminary step for backwards particle tracking. Tentative next step is to introduce an eager read capability that builds an index in advance and lets you read in either direction. Idea is to keep reader types interchangeable so consumers can use the lazy forward pattern or the eager bidirectional pattern. That will come next, this is a pure refactor to lay groundwork.