refactor: introduce binary file header type (#2389)
Followup to #2384. Collect header properties into a type and expose that from the binary file reader instead of properties individually. Just a tidying step. Next, to come separately, also optionally (?) as it's a bigger change with performance considerations, scan headers and build an index, which is needed for direction-agnostic readers

The header's file position property pos is unused for now. Also the "next" header is awkward, peeking at the next kper/kstp leaves the rest of it unfilled, but it can be removed when an index is in as then there's no need to look ahead.