fix(prt-prp): fix duplicate entries in memory manager (#2422)
The PRT PRP package was adding duplicate entries to the memory manager, producing warning messages in simulation output. The BndType's allocate_arrays() was called from prp_ar() instead of from PRP's override of allocate_arrays(). This has been the case for a while, but only came to light recently when the memory manager began using a hash map.