refactor(prt): reference/rename max poly verts constant (#2470)
Noticed in review of #2446, MAX_POLY_CELLS was not used everywhere, the value (10) was hardcoded in a few places. This is the maximum number of vertices (also lateral faces) in a cell polygon so rename it MAX_POLY_VERTS, and reference it where needed.

Also noticed 2 routines no longer used (now replaced by cell conversion routines in CellUtil.f90), remove them.