fix(prt): terminate at assigned boundary faces (#2475)
Previously the particle tracking method was aware of boundary cell flows but not boundary faces. When internal boundary faces were assigned, i.e. an IFLOWFACE is adjacent to another active cell, this could lead to cycles, as the two adjacent cells failed to agree on the flow through their shared face.

Terminate particles exiting a cell through a boundary face, internal or external. An option may be added later to toggle between this and MODPATH behavior (ignoring internal boundary faces).