fix(prt): consistent behavior at top of partly saturated cells (#2496)
Discussion has been had regarding behavior at the water table. This may be refined in future.

Behavior just introduced in #2478 and #2495 keeps particles immobile/active at the water table if they reach it from below with DRY_TRACKING_METHOD DROP or STAY, while with STOP particles would terminate on reaching the water table. This is in hindsight overfitting to the expected Keating result at the expense of design aims

It seems more correct to interpret DRY_TRACKING_METHOD as only applying to particles which are dry after the flow solution is available, before the particle track is solved for the time step. Trying to find distinct interpretations in various contexts within the tracking method, e.g. if a particle obtains a valid exit solution through the "top" (i.e. water table) of a partially saturated cell, could explode the scope of the dry tracking method option so seems best avoided

With this, particles terminate if they reach the water table from below. This is consistent with MP7. DRY_TRACKING_METHOD applies only if the water table drops below them during the simulation, or if particles are released above the water table into a Newton flow system.

This does imply accepting some particles will stop early in Keating for some compilers and optimization options; numerical noise in the flow solution will try to send some of them back upwards after they drop to the water table, so they fall straight to water and terminate. Most will keep doing the "waterfall"/cascade down the sides of the perched aquifer as we expect. Maybe the flow system can be tweaked to fix the noisy flows and get them all to cooperate?

Separately, recent commits got a face numbering calculation in MethodCell.f90 wrong in the general case, but in such a way as to happen to work for the rectilinear case

Also miscellaneous cleanup