docs(dfns): deprecate reader, infer from other attributes (#2565)
The reader attribute for DFN variables seems unnecessary. Deprecate it and infer it in mf6ivar.py from shape/type. It is only used for generating documentation. Leave reader in the DFNs for now in case other people are using it. Also, the DFN spec listed old, outdated values from MF2005 like "u1ddbl" as still in use, clean that up.