ci: fall back to ifort 2021.6 on macos-13 (#2585)
As mentioned in #2575 we lost ifort 2021.7 on macOS, this is a temporary stopgap til we move to ifx.

Now is the time to drop the macos-13 (intel mac) distribution entirely though. To keep it after the move to ifx we'd have to convert it to gfortran like the apple silicon distribution, and in any case github will retire macos-13 soon anyway.