fix(prt): balance budget, add termination term (#2452)
Particle mass previously remained in storage regardless of particle status.

Add a termination budget term and (for now) assign all terminating particle mass to it. This balances the cumulative storage budget. This is first in a sequence of budget reworks. Next planned is an analog to transport's SSM so that mass terminating at assigned boundaries can be reported for the boundary as done for transport models. After that, advanced packages in the same pattern. But even then, a term for non-boundary termination will still be needed.

Also, report the cell-by-cell storage term, only the cumulative was reported up to now.

Incidentally, add a function to get a particle's unique ID (composite key of model ID, PRP ID, release point ID, and release time) as a string. We had this before, and removed it as it went unused, but it's convenient to use here as a hash table key to track which particles have terminated.

This includes a new test focused on the cumulative and cell by cell budget reporting, and renames an existing test which was called "budget" but really focused on intercell mass flows.