ci(update): replace macos-13 runner with macos-15-intel (#2604)
Seems like the macos-15-intel image will be available until mid-2027 which relieves a bit of near term pressure to migrate to ifx, but we still have no guarantee that intel will keep the old oneapi installers available so ifx migration is still a priority. But this at least lets us keep CI testing intel compilers on mac for now.