chore(dependencies): add xugrid to pixi environment (#2611)
So we don't have to install separately in CI, where previously we got it from PyPI. This avoids CI failures due to the macos-15-intel image lacking LLVM build files. We hit this because xugrid depends on numba which depends on llvmlite, and llvmlite doesn't have a wheel for intel macos 15.