fix(prt-prp): fix release local z coord calculation (#2624)
Fix conversion of local z release point coordinates to model z coordinates: take into account whether the cell is convertible or confined, and avoid incorrectly computing model z coordinates to be below the cell bottom. If convertible, define the cell's vertical extent from the bottom to the saturated top, clamping sat top to the bottom if head is below the cell bottom elevation; if confined, use the cell's full vertical extent from the bottom to the geometric top.

In Newton models, the failure to clamp to the bottom could produce pathlines with invalid (negative) times. A particle released by local z coord into a cell with head below the cell bottom would effectively be placed below the cell bottom, which poisoned subsequent tracking calculations.

The user must now provide a binary grid file to PRT FMI if release coordinates are configured with LOCAL_Z, to distinguish convertible from confined cells at release time.

Internally, this required giving FMI a pointer to cell type as configured in GWF NPF. If GWF and PRT are in the same simulation, cell type is set from the EXG. If in separate sims, it comes from the GRB file.

This in turn required tweaking the grid file reader, as the preexisting array read functions did not support reading into preallocated, memory-managed arrays.