docs(prt): distinguish nonexistent from inactive cells (#2625)
PRT docs and error messages previously failed to distinguish between cells excluded from the model grid with IDOMAIN and cells merely inactive due e.g. to drying. These two concepts are distinct and should not be confused: PRT will not permit a particle to be released into a cell with IDOMAIN = 0 (excluded from the simulation) or -1 (vertical pass through), but it will permit release into an inactive cell. Behavior in this case is determined by factors such as whether the DRAPE option is activated, whether the flow model uses the Newton formulation, and (if so) which DRY_TRACKING_METHOD is configured.